[test] Add util/GetArchivesize test
